requirements.txt
asgiref==3.4.1
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-nested-routers==0.93.3
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
